+++ include/settings.h
@@ -29,13 +29,15 @@
 // Path to the latest Firmware
 #define FW_BINFILE "firmware.bin"
 
-// Path to the latest filesystem
-#define FW_SPIFFSFILE "spiffs.bin"
+//Data directory Source
+#define DATA_UPDATEURL "https://raw.githubusercontent.com/doudar/SmartSpin2k/LittleFS/data"
 
-// name of local file to save configuration in SPIFFS
+#define DATA_FILELIST "/list.json"
+
+// name of local file to save configuration in LittleFS
 #define configFILENAME "/config.txt"
 
-// name of local file to save Physical Working Capacity in Spiffs
+// name of local file to save Physical Working Capacity in LittleFS
 #define userPWCFILENAME "/userPWC.txt"
 
 // Default Incline Multiplier.