+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
  - updated CA for OTA updates
+ - Converted filesystem from SPIFFS to LittleFS
 
 
 ### Hardware